* Renane tab to `Server Configuration Parameters` * Generate argocd-util command docs
