+++ docs/operator-manual/server-commands/argocd-util_projects.md
@@ -1,0 +1,25 @@
+## argocd-util projects
+
+
+
+### Synopsis
+
+
+
+```
+argocd-util projects [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for projects
+```
+
+### SEE ALSO
+
+* [argocd-util](argocd-util.md)	 - argocd-util has internal tools used by Argo CD
+* [argocd-util projects generate-allow-list](argocd-util_projects_generate-allow-list.md)	 - Generates project allow list from the specified clusterRole file
+* [argocd-util projects update-role-policy](argocd-util_projects_update-role-policy.md)	 - Implement bulk project role update. Useful to back-fill existing project policies or remove obsolete actions.
+
+###### Auto generated by spf13/cobra on 9-Nov-2020
